some reformats to the code
