+++ test/elektra_test.cpp
@@ -1,4 +1,5 @@
 #include "gtest/gtest.h"
+#include "tests/test_parallel_batch_connected.h"
 #include "tests/test_parallel_euler_tour_tree.h"
 #include "tests/test_parallel_skip_list.h"
